refactor: InputValidation에 있던 로또번호 체그 로직 LottoMachine로 이동
