+++ src/test/java/domain/LottoMachineTest.java
@@ -6,8 +6,7 @@ import org.junit.jupiter.api.Test;
 
 import java.util.*;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.Assertions.*;
 
 public class LottoMachineTest {
 
@@ -18,6 +17,59 @@ public class LottoMachineTest {
         LottoMachine lottoMachine = new LottoMachine(givenMoney);
         //when //than
         assertThat(lottoMachine.getLottoTicketCount()).isEqualTo(14);
+    }
+    @Test
+    void _1에서_45_사이의_로또번호와_중복된_보너스번호_입력_하면_에러_리턴() {
+        //given
+        String bonusNumbers = "45";
+        List<Integer> winningNumber = new ArrayList<>();
+        winningNumber.add(1);
+        winningNumber.add(2);
+        winningNumber.add(3);
+        winningNumber.add(4);
+        winningNumber.add(5);
+        winningNumber.add(45);
+        //when
+        LottoMachine machine = new LottoMachine("1000");
+        //then
+        assertThatThrownBy(
+                () -> machine.getBonusBall(bonusNumbers, winningNumber))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+    @Test
+    @DisplayName("입력 숫자의 범위가 1-45가 아닐 경우 예외가 발생한다.")
+    public void checkBound() {
+        //given
+        String bonusNumbers = "46";
+        List<Integer> winningNumber = new ArrayList<>();
+        winningNumber.add(1);
+        winningNumber.add(2);
+        winningNumber.add(3);
+        winningNumber.add(4);
+        winningNumber.add(5);
+        winningNumber.add(45);
+
+        //when
+        LottoMachine machine = new LottoMachine("1000");
+        // then
+        assertThatIllegalArgumentException().isThrownBy(() -> machine.getBonusBall(bonusNumbers, winningNumber));
+    }
+    @Test
+    void _1에서_45_사이의_로또보너스번호_입력_하면_숫자로_리턴() {
+        //given
+        String bonusNumbers = "45";
+        List<Integer> winningNumber = new ArrayList<>();
+        winningNumber.add(1);
+        winningNumber.add(2);
+        winningNumber.add(3);
+        winningNumber.add(4);
+        winningNumber.add(5);
+        winningNumber.add(6);
+        //when
+        LottoMachine machine = new LottoMachine("1000");
+        int result =  machine.getBonusBall(bonusNumbers, winningNumber);
+        //then
+        assertThat(result).isEqualTo(45);
     }
     @Test
     void 문자_100을_입력받으면_런타임에러를_리턴한다() {
@@ -40,6 +92,7 @@ public class LottoMachineTest {
                 new LottoMachine(givenMoney)).
                 isInstanceOf(RuntimeException.class);
     }
+
     @Test
     void 문자_음수1000_입력받으면_런타임에러를_리턴한다() {
         //given
